Fixed few structures and named unks.

+++ ClientDefinitions/DBC/Cataclysm/LiquidObjectEntry.cs
@@ -5,9 +5,9 @@ namespace FileStructures.DBC.Cataclysm
     public sealed class LiquidObjectEntry
     {
         public uint Id;
-        public uint Unk0;
-        public uint Unk1;
-        public uint Unk2;
+        public float Unk0;
+        public float Unk1;
+        public int Unk2;
         public uint Unk3;
         public uint Unk4;
     }
